* develop:
  Verion typo fix
  ## [1.5.4] - 2022-09-14 ## Added - Fully refactored frontend ui - Gradient generator for tailwind config - Factory generation for the CRUD ## Fixed - Duplicated request validation - Removed file field require from update request - In the file relation domain layer resource the badly generated relation name now points the correct relation
  Renamed to more relevant field selector names
  Regenartor dir fix
  Gradient generator
  Added Dahboard generation Fixed small errors
  Added file and picture upload handling Notification updates Small tweaks
  Use common response key for success and failure: message instead of failed
  Fixings
  Factory fixed, useModel route name fixed routeBlock fixed added relation to api controller
  Add foreign generation to crudMenu
  Fix component generation Add missing column prop to ApiMultiselect
  Repgenerator will not generate composable base column config
  Added page generation for nuxt Added factory generation alpha version
  Added column generation Added larafetch generation
  Revert
  Fix edit stub
  changelog update fix request validation generation error
  "@tailwindcss/forms": "^0.5.2"
  ## [1.5.3] - 2022-09-09 ## Fixed - Clear up frontend depedency tree

+++ src/Domain/Pattern/Services/RepgeneratorStaticFilesService.php
@@ -14,6 +14,34 @@ class RepgeneratorStaticFilesService
 
     }
 
+    public function copyFilesToGivenLocation(array $files, string $framework, string $folderPath  = 'js/Abstraction/'): array
+    {
+        $generatedFiles = [];
+        foreach ($files as $fileOriginal) {
+            $fileParts = explode('/', $fileOriginal);
+            foreach ( $fileParts as $index => $filePart ) {
+                if ( end($fileParts) ==  $filePart ) {
+                    continue;
+                }
+                $partsSoFar = implode('/',array_slice($fileParts,0, $index+1));
+                if ( !is_dir(resource_path($folderPath . $partsSoFar))) {
+                    mkdir(resource_path($folderPath . $partsSoFar), 0777, true);
+                }
+            }
+            $nameWithExtension = end($fileParts);
+            $file = $this->getFrontendStatic($fileOriginal);
+
+            $frontendReplacer = app(RepgeneratorFrontendFrameworkHandlerService::class);
+            $file = $frontendReplacer->replaceForFramework($framework, $file);
+
+            if (!file_exists($path = resource_path($folderPath. $fileOriginal))) {
+                file_put_contents($path, $file);
+                $generatedFiles[] = new RepgeneratorStaticFileAdapter($nameWithExtension, $path);
+            }
+        }
+        return $generatedFiles;
+    }
+
     public function copyStaticFrontendFiles(string $framework, string $folderPath = 'js/Abstraction/'): array {
         $files = [
             "components/DataTable/ColumnHeader.vue",
@@ -34,7 +62,7 @@ class RepgeneratorStaticFilesService
             "components/FileUpload.vue",
             "composables/model.js",
             "composables/useNotifications.ts",
-            'utils/$larafetch.ts',
+            "router/router.js"
         ];
 
         $generatedFiles = [];
@@ -55,11 +83,10 @@ class RepgeneratorStaticFilesService
             $frontendReplacer = app(RepgeneratorFrontendFrameworkHandlerService::class);
             $file = $frontendReplacer->replaceForFramework($framework, $file);
 
-            //if (!file_exists($path = app_path($fileOriginal))) {
-            $path = resource_path($folderPath. $fileOriginal);
+            if (!file_exists($path = resource_path($folderPath. $fileOriginal))) {
             file_put_contents($path, $file);
             $generatedFiles[] = new RepgeneratorStaticFileAdapter($nameWithExtension, $path);
-            //}
+            }
         }
         return $generatedFiles;
     }
@@ -98,12 +125,11 @@ class RepgeneratorStaticFilesService
             $name = explode('.', $nameWithExtension)[0];
             $file = $this->getStatic($name);
 
-            //if (!file_exists($path = app_path($fileOriginal))) {
-                $path = app_path($fileOriginal);
+            if (!file_exists($path = app_path($fileOriginal))) {
                 file_put_contents($path, $file);
 
                 $generatedFiles[] = new RepgeneratorStaticFileAdapter($nameWithExtension, $path);
-            //}
+            }
         }
         return $generatedFiles;
     }
